# docs/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files": {
      "get": {
        "operationId": "FileController_list",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      },
      "post": {
        "operationId": "FileController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
...
    "/files/{id}": {
      "get": {
        "operationId": "FileController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      },
      "delete": {
        "operationId": "FileController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "File"
        ]
      }
    },
...
    "/assets/{key}": {
      "get": {
        "operationId": "AssetController_get",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Asset"
        ]
      }
    },
    "/_ipx/cache/{sourcePath}": {
      "delete": {
        "operationId": "IpxCacheController_invalidate",
        "parameters": [
          {
            "name": "sourcePath",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IpxCache"
        ]
      }
    },
...
  },
  "info": {
    "title": "nest-server",
    "description": "Template-fähiger NestJS-Server",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
...
  }
}